refactor(digitize): restore all inline comments + move import/export to admin.py
